pyproject.toml
[project]
name = "succuba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Josh Klein", email = "gigacoderx@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-cloud-texttospeech>=2.34.0",
    "moviepy>=2.2.1",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
start-tts = "succuba.__main__:main"
